moved list to own file, + more refactor
