+++ easyCheck.h
@@ -1,7 +1,11 @@
 #ifndef EASYCHECK_H
 #define EASYCHECK_H
 
+#include <netinet/in.h>
+#include <semaphore.h>
+
 #include "errorHelpers.h"
+#include "options.h"
 
 // error checked and retried (if EINTR) versions of some message queue functions
 #define mq_send_(mqdes, msg_ptr, msg_len, msg_prio, ...)\
@@ -148,5 +152,98 @@
 // extra:
 #define open_(path, oflag, ...)\
 	CHECK_RETRY(open(path, oflag, ##__VA_ARGS__))
+
+sigset_t MY_WARN_UNUSED
+make_sigset_t_(int signo) {
+	sigset_t sigset;
+	ERR_NEG1(sigemptyset(&sigset));
+
+	// to allow returning emptyset if signo == 0
+	if (signo) {
+		sigaddset_(&sigset, signo);
+	}
+
+	return sigset;
+}
+// bigbrane """vararg""" lul
+#define make_sigset_t_0_ARGS() make_sigset_t_(0)
+#define make_sigset_t_1_ARG(signo) make_sigset_t_(signo)
+#define make_sigset_t_X_ARGS(x, A, FUNC, ...) FUNC
+#define make_sigset_t(...) make_sigset_t_X_ARGS(, ##__VA_ARGS__,\
+		make_sigset_t_1_ARG(__VA_ARGS__),\
+		make_sigset_t_0_ARGS(__VA_ARGS__))
+
+#define sigset_make(...) make_sigset_t(__VA_ARGS__)
+// gonna start adding ^these b/c the POSIX ones look like this
+// will keep the old format too tho b/c I have some old programs with them :(
+// we do sum backwards compatibility ig
+
+pthread_mutex_t MY_WARN_UNUSED
+make_pthread_mutex_t_(const pthread_mutexattr_t* attr) {
+	pthread_mutex_t mutex;
+	pthread_mutex_init_(&mutex, attr);
+	return mutex;
+}
+#define make_pthread_mutex_t_0_ARGS() make_pthread_mutex_t_(NULL)
+#define make_pthread_mutex_t_1_ARG(attr) make_pthread_mutex_t_(attr)
+#define make_pthread_mutex_t_X_ARGS(x, A, FUNC, ...) FUNC
+#define make_pthread_mutex_t(...) make_pthread_mutex_t_X_ARGS(, ##__VA_ARGS__,\
+		make_pthread_mutex_t_1_ARG(__VA_ARGS__),\
+		make_pthread_mutex_t_0_ARGS(__VA_ARGS__))
+#define pthread_mutex_make(...) make_pthread_mutex_t(__VA_ARGS__)
+
+pthread_attr_t MY_WARN_UNUSED
+make_pthread_attr_t(void) {
+	pthread_attr_t attr;
+	pthread_attr_init_(&attr);
+	return attr;
+}
+#define pthread_attr_make() make_pthread_attr_t()
+
+pthread_cond_t MY_WARN_UNUSED
+make_pthread_cond_t_(const pthread_condattr_t* attr) {
+	pthread_cond_t cond;
+	pthread_cond_init_(&cond, attr);
+	return cond;
+}
+#define make_pthread_cond_t_0_ARGS() make_pthread_cond_t_(NULL)
+#define make_pthread_cond_t_1_ARG(attr) make_pthread_cond_t_(attr)
+#define make_pthread_cond_t_X_ARGS(x, A, FUNC, ...) FUNC
+#define make_pthread_cond_t(...) make_pthread_cond_t_X_ARGS(, ##__VA_ARGS__,\
+		make_pthread_cond_t_1_ARG(__VA_ARGS__),\
+		make_pthread_cond_t_0_ARGS(__VA_ARGS__))
+#define pthread_cond_make(...) make_pthread_cond_t(__VA_ARGS__)
+
+pthread_barrier_t MY_WARN_UNUSED
+make_pthread_barrier_t_(const pthread_barrierattr_t* attr, unsigned count) {
+	pthread_barrier_t barrier;
+	pthread_barrier_init_(&barrier, attr, count);
+	return barrier;
+}
+#define make_pthread_barrier_t_0_ARGS() make_pthread_barrier_t_(NULL, 2)
+#define make_pthread_barrier_t_1_ARG(count) make_pthread_barrier_t_(NULL, count)
+#define make_pthread_barrier_t_2_ARGS(attr, count) make_pthread_barrier_t_(attr, count)
+#define make_pthread_barrier_t_X_ARGS(x, A, B, FUNC, ...) FUNC
+#define make_pthread_barrier_t(...) make_pthread_barrier_t_X_ARGS(, ##__VA_ARGS__,\
+	make_pthread_barrier_t_2_ARGS(__VA_ARGS__),\
+	make_pthread_barrier_t_1_ARG(__VA_ARGS__),\
+	make_pthread_barrier_t_0_ARGS(__VA_ARGS__))
+#define pthread_barrier_make(...) make_pthread_barrier_t(__VA_ARGS__)
+
+sem_t MY_WARN_UNUSED
+make_sem_t_(int pshared, unsigned value) {
+	sem_t sem;
+	sem_init_(&sem, pshared, value);
+	return sem;
+}
+#define make_sem_t_0_ARGS() make_sem_t_(0, 1) /* binary sem by default ig lol */
+#define make_sem_t_1_ARG(count) make_sem_t_(0, count)
+#define make_sem_t_2_ARGS(pshared, count) make_sem_t_(pshared, count)
+#define make_sem_t_X_ARGS(x, A, B, FUNC, ...) FUNC
+#define make_sem_t(...) make_sem_t_X_ARGS(, ##__VA_ARGS__,\
+	make_sem_t_2_ARGS(__VA_ARGS__),\
+	make_sem_t_1_ARG(__VA_ARGS__),\
+	make_sem_t_0_ARGS(__VA_ARGS__))
+#define sem_make(...) make_sem_t(__VA_ARGS__)
 
 #endif // EASYCHECK_H
