+++ katwikOpsys.h
@@ -52,7 +52,7 @@
 #include "colors.h"
 
 // extra attributes
-#include "extraAttr.h"
+#include "easyAttr.h"
 
 // smol things
 #include "littleHelpers.h"
@@ -60,18 +60,14 @@
 // error stuff
 #include "errorHelpers.h"
 
+// functions that are already error checked
 #include "easyCheck.h"
+
+// homebrew linked list xdd
+#include "myList.h"
 
 // long story short, block comments don't nest nicely lol
 #define TEMP_REMOVE 0
-
-// version of DBGprintf that only compiles in DEBUG_MODE
-// kinda want DBGprintf to be nice and accesible normally
-#if DEBUG_MODE
-#define DBGonly(...) DBGprintf(__VA_ARGS__)
-#else // DEBUG_MODE
-#define DBGonly(...)
-#endif // DEBUG_MODE
 
 // bit for functions that sometimes error depending on settings lol
 // TODO: generalize RETURN_FAIL? idk
@@ -556,298 +552,6 @@ MIGHT_ERR() MY_NON_NULL(5)
 		RETURN_OKAY;
 	}
 
-typedef struct MyNode_ {
-	LIST_TYPE val;
-	struct MyNode_ *next;
-} MyNode;
-
-typedef struct MyList_ {
-	MyNode *head,
-	       *tail; // okay so let's get real
-	// the only real reason to have a tail pointer in a singly
-	// linked list is to have fast operations for the ends, but I
-	// never used this lol, probably TODO: remove this lol
-} MyList;
-
-MyNode* MY_WARN_UNUSED
-newMyNode (LIST_TYPE val) {
-	MyNode *newMyNode;
-	ERR_NULL( newMyNode = malloc(sizeof(MyNode)) );
-
-	newMyNode->val = (LIST_TYPE) val;
-	newMyNode->next = NULL;
-	return newMyNode;
-}
-
-MyList* MY_WARN_UNUSED
-newMyList (void) {
-	MyList* list;
-
-	// calloc to correctly set NULL pointers
-	ERR_NULL( list = calloc(1, sizeof(MyList)) );
-	return list;
-}
-
-MIGHT_ERR() MY_NON_NULL(1, 2) MY_FORMAT(printf, 2, 0)
-	printMyList(const MyList* const list, const char* restrict format) {
-		NON_ATTR_IF(!list) {
-			RETURN_FAIL("bruh where's the list");
-		}
-
-		DBGonly("list:%p format:%s\n[H:%p T:%p] ",
-				(void*) list, format,
-				(void*) list->head, (void*) list->tail
-		       );
-
-		for (MyNode* node = list->head; node; node = node->next) {
-			printf_(format, node->val);
-		}
-
-		printf_("\n");
-
-		RETURN_OKAY;
-	}
-
-size_t MY_NON_NULL(1) MY_WARN_UNUSED
-myListLength(const MyList* const list) {
-	size_t length = 0;
-	for (MyNode* node = list->head; node; node = node->next) {
-		++length;
-	}
-
-	return length;
-}
-
-MyNode* MY_WARN_UNUSED MY_NON_NULL(1)
-	popAt(MyList* const list, size_t pos) {
-		MyNode *prev = NULL, *node = list->head;
-
-		// is this redundant btw? TODO: remove and check lol
-		if (!list->head) {
-			return NULL; // Can't delete from empty list
-		}
-
-		for (size_t i = 0; i < pos; ++i) {
-			if (node) {
-				prev = node;
-				node = node->next;
-			} else {
-				return NULL; // Can't delete outside of list
-			}
-		}
-
-		// idk there's a few SIGSEGVs without this lol
-		// it's getting pretty messy now, I should rewrite this whole function tbh
-		if (!node) {
-			return NULL;
-		}
-
-		// If there's a previous node relink, otherwise we're deleting the head node, so update that
-		prev ?
-			(prev->next = node->next) :
-			(list->head = node->next);
-
-		// If we're deleting the tail node, update that
-		if (!node->next) {
-			list->tail = prev;
-		}
-
-		// clean up and return
-		node->next = NULL;
-		return node;
-	}
-
-MyNode* MY_WARN_UNUSED MY_NON_NULL(1)
-	peekNodeAt(const MyList* const list, size_t pos) {
-		MyNode* node = list->head;
-
-		if (!list->head) {
-			return NULL; // peek empty list
-		}
-
-		for (size_t i = 0; i < pos; ++i) {
-			if (node) {
-				node = node->next;
-			} else {
-				return NULL; // peek outside list
-			}
-		}
-
-		return node;
-	}
-
-LIST_TYPE MY_WARN_UNUSED MY_NON_NULL(1)
-	peekValAt(const MyList* const list, size_t pos) {
-		MyNode* node;
-
-		// TODO: make this fancy or something lol
-		if (NULL == (node = peekNodeAt(list, pos)) ) {
-			errno = MY_ERRNO;
-			ERR("Can't peek value outside list");
-		}
-
-		return node->val;
-	}
-
-LIST_TYPE MY_WARN_UNUSED MY_NON_NULL(1)
-	popValAt(MyList* const list, size_t pos) {
-		MyNode* node;
-		LIST_TYPE val;
-
-		if (NULL == (node = popAt(list, pos)) ) {
-			errno = MY_ERRNO;
-			ERR("Can't pop value outside list"); // ?
-		}
-
-		val = node->val;
-		FREE(node); // we do a little leakn't
-
-		return val;
-	}
-
-LIST_TYPE MY_WARN_UNUSED MY_NON_NULL(1)
-	peekFirstVal(const MyList* const list) {
-		return peekValAt(list, 0);
-	}
-
-LIST_TYPE MY_WARN_UNUSED MY_NON_NULL(1)
-	peekLastVal(const MyList* const list) {
-		size_t sz = myListLength(list);
-		return peekValAt(list, sz - 1);
-	}
-
-LIST_TYPE MY_WARN_UNUSED MY_NON_NULL(1)
-	popFirstVal(MyList* const list) {
-		return popValAt(list, 0);
-	}
-
-LIST_TYPE MY_WARN_UNUSED MY_NON_NULL(1)
-	popLastVal(MyList* const list) {
-		size_t sz = myListLength(list);
-		return popValAt(list, sz - 1);
-	}
-
-MyNode* MY_WARN_UNUSED MY_NON_NULL(1)
-	popLast(MyList* const list) {
-		size_t sz = myListLength(list);
-		return popAt(list, sz - 1);
-	}
-
-MyNode* MY_WARN_UNUSED MY_NON_NULL(1)
-	popFirst(MyList* const list) {
-		return popAt(list, 0);
-	}
-
-MyList* MY_NON_NULL(1)
-	clearMyList(MyList* const list) {
-		for (MyNode* first; (first = popFirst(list));) {
-			FREE(first);
-		}
-
-		return list;
-	}
-
-void MY_NON_NULL(1)
-	deleteMyList (MyList* list) {
-		clearMyList(list);
-		FREE(list);
-	}
-
-MyList* MY_NON_NULL(1)
-	deleteAt(MyList* const list, size_t pos) {
-		MyNode* node = popAt(list, pos);
-		FREE(node);
-		return list;
-	}
-
-MyList* MY_NON_NULL(1)
-	deleteLast(MyList* const list) {
-		MyNode* node = popLast(list);
-		FREE(node);
-		return list;
-	}
-
-MyList* MY_NON_NULL(1)
-	deleteFirst(MyList* const list) {
-		MyNode* node = popAt(list, 0);
-		FREE(node);
-		return list;
-	}
-
-MyList* MY_NON_NULL(1, 3)
-	insertAfter(MyList* const list, size_t pos, MyNode* const newMyNode) {
-		MyNode *node = list->head;
-
-		for (size_t i = 0; i < pos; ++i) { // Find the node we're inserting after
-			if (node && node->next) { // Keep moving if this is a node and there's a next one
-				node = node->next;
-			} else {
-				break; // will by default insert after the last one
-			}
-		}
-
-		node == list->tail ? (list->tail = newMyNode) : (newMyNode->next = node->next);
-
-		// Be careful in case we insertAfter in an empty list
-		!node ? (list->head = newMyNode) : (node->next = newMyNode);
-		return list;
-	}
-
-MyList* MY_NON_NULL(1, 3)
-	insertBefore(MyList* const list, size_t pos, MyNode* const newMyNode) {
-		MyNode *prev = NULL, *node = list->head;
-
-		for (size_t i = 0; i < pos; ++i) { // Find the node we're inserting before
-			if (node && node->next) {
-				prev = node;
-				node = node->next;
-			} else {
-				break; // will by default insert before the last one
-			}
-		}
-
-		node == list->head ? (list->head = newMyNode) : (prev->next = newMyNode);
-
-		// Be careful in case we insertBefore in an empty list
-		!node ? (list->tail = newMyNode) : (newMyNode->next = node);
-		return list;
-	}
-
-MyList* MY_NON_NULL(1)
-	insertValAfter(MyList* const list, size_t pos, LIST_TYPE newVal) {
-		// NB: returns list after allocation then insertion
-		return insertAfter(list, pos, newMyNode(newVal));
-	}
-
-MyList* MY_NON_NULL(1)
-	insertValBefore(MyList* const list, size_t pos, LIST_TYPE newVal) {
-		return insertBefore(list, pos, newMyNode(newVal));
-	}
-
-MyList* MY_NON_NULL(1, 2)
-	insertLast(MyList* const list, MyNode* const node) {
-		size_t sz = myListLength(list);
-		insertAfter(list, sz - 1, node);
-		return list;
-	}
-
-MyList* MY_NON_NULL(1, 2)
-	insertFirst(MyList* const list, MyNode* const node) {
-		insertBefore(list, 0, node);
-		return list;
-	}
-MyList* MY_NON_NULL(1)
-	insertValLast(MyList* const list, LIST_TYPE newVal) {
-		return insertLast(list, newMyNode(newVal));
-	}
-
-MyList* MY_NON_NULL(1)
-	insertValFirst(MyList* const list, LIST_TYPE newVal) {
-		return insertFirst(list, newMyNode(newVal));
-	}
-
-// socket stuff will start going below this point
-
 MIGHT_ERR() MY_NON_NULL(1)
 	init_sockaddr_in(struct sockaddr_in* addr,
 			sa_family_t family, in_port_t port, uint32_t s_addr) {
@@ -885,104 +589,12 @@ in_addr_t MY_NON_NULL(1)
 		return addr;
 	}
 
+// easy way to set a signal handler AND error check
 void sethandler(void (*signalHandler) (int), int sigNo) {
 	struct sigaction action = {0};
 	action.sa_handler = signalHandler;
 	ERR_NEG1( sigaction(sigNo, &action, NULL) ); // TODO: change this
 }
-
-sigset_t MY_WARN_UNUSED
-make_sigset_t_(int signo) {
-	sigset_t sigset;
-	ERR_NEG1(sigemptyset(&sigset));
-
-	// to allow returning emptyset if signo == 0
-	if (signo) {
-		sigaddset_(&sigset, signo);
-	}
-
-	return sigset;
-}
-// bigbrane """vararg""" lul
-#define make_sigset_t_0_ARGS() make_sigset_t_(0)
-#define make_sigset_t_1_ARG(signo) make_sigset_t_(signo)
-#define make_sigset_t_X_ARGS(x, A, FUNC, ...) FUNC
-#define make_sigset_t(...) make_sigset_t_X_ARGS(, ##__VA_ARGS__,\
-		make_sigset_t_1_ARG(__VA_ARGS__),\
-		make_sigset_t_0_ARGS(__VA_ARGS__))
-
-#define sigset_make(...) make_sigset_t(__VA_ARGS__)
-// gonna start adding ^these b/c the POSIX ones look like this
-// will keep the old format too tho b/c I have some old programs with them :(
-// we do sum backwards compatibility ig
-
-pthread_mutex_t MY_WARN_UNUSED
-make_pthread_mutex_t_(const pthread_mutexattr_t* attr) {
-	pthread_mutex_t mutex;
-	pthread_mutex_init_(&mutex, attr);
-	return mutex;
-}
-#define make_pthread_mutex_t_0_ARGS() make_pthread_mutex_t_(NULL)
-#define make_pthread_mutex_t_1_ARG(attr) make_pthread_mutex_t_(attr)
-#define make_pthread_mutex_t_X_ARGS(x, A, FUNC, ...) FUNC
-#define make_pthread_mutex_t(...) make_pthread_mutex_t_X_ARGS(, ##__VA_ARGS__,\
-		make_pthread_mutex_t_1_ARG(__VA_ARGS__),\
-		make_pthread_mutex_t_0_ARGS(__VA_ARGS__))
-#define pthread_mutex_make(...) make_pthread_mutex_t(__VA_ARGS__)
-
-pthread_attr_t MY_WARN_UNUSED
-make_pthread_attr_t(void) {
-	pthread_attr_t attr;
-	pthread_attr_init_(&attr);
-	return attr;
-}
-#define pthread_attr_make() make_pthread_attr_t()
-
-pthread_cond_t MY_WARN_UNUSED
-make_pthread_cond_t_(const pthread_condattr_t* attr) {
-	pthread_cond_t cond;
-	pthread_cond_init_(&cond, attr);
-	return cond;
-}
-#define make_pthread_cond_t_0_ARGS() make_pthread_cond_t_(NULL)
-#define make_pthread_cond_t_1_ARG(attr) make_pthread_cond_t_(attr)
-#define make_pthread_cond_t_X_ARGS(x, A, FUNC, ...) FUNC
-#define make_pthread_cond_t(...) make_pthread_cond_t_X_ARGS(, ##__VA_ARGS__,\
-		make_pthread_cond_t_1_ARG(__VA_ARGS__),\
-		make_pthread_cond_t_0_ARGS(__VA_ARGS__))
-#define pthread_cond_make(...) make_pthread_cond_t(__VA_ARGS__)
-
-pthread_barrier_t MY_WARN_UNUSED
-make_pthread_barrier_t_(const pthread_barrierattr_t* attr, unsigned count) {
-	pthread_barrier_t barrier;
-	pthread_barrier_init_(&barrier, attr, count);
-	return barrier;
-}
-#define make_pthread_barrier_t_0_ARGS() make_pthread_barrier_t_(NULL, 2)
-#define make_pthread_barrier_t_1_ARG(count) make_pthread_barrier_t_(NULL, count)
-#define make_pthread_barrier_t_2_ARGS(attr, count) make_pthread_barrier_t_(attr, count)
-#define make_pthread_barrier_t_X_ARGS(x, A, B, FUNC, ...) FUNC
-#define make_pthread_barrier_t(...) make_pthread_barrier_t_X_ARGS(, ##__VA_ARGS__,\
-	make_pthread_barrier_t_2_ARGS(__VA_ARGS__),\
-	make_pthread_barrier_t_1_ARG(__VA_ARGS__),\
-	make_pthread_barrier_t_0_ARGS(__VA_ARGS__))
-#define pthread_barrier_make(...) make_pthread_barrier_t(__VA_ARGS__)
-
-sem_t MY_WARN_UNUSED
-make_sem_t_(int pshared, unsigned value) {
-	sem_t sem;
-	sem_init_(&sem, pshared, value);
-	return sem;
-}
-#define make_sem_t_0_ARGS() make_sem_t_(0, 1) /* binary sem by default ig lol */
-#define make_sem_t_1_ARG(count) make_sem_t_(0, count)
-#define make_sem_t_2_ARGS(pshared, count) make_sem_t_(pshared, count)
-#define make_sem_t_X_ARGS(x, A, B, FUNC, ...) FUNC
-#define make_sem_t(...) make_sem_t_X_ARGS(, ##__VA_ARGS__,\
-	make_sem_t_2_ARGS(__VA_ARGS__),\
-	make_sem_t_1_ARG(__VA_ARGS__),\
-	make_sem_t_0_ARGS(__VA_ARGS__))
-#define sem_make(...) make_sem_t(__VA_ARGS__)
 
 #endif // HEADER_OKAY
 #undef HEADER_OKAY
